Fix: appliquer les fix demandés, fix le modal/fix Renamer
